drivers/staging: ion: Use unsigned long instead of int
It just doesnt fit otherwise and needs to return int/long to avoid warnings

+++ drivers/staging/android/ion/mtk/ion_drv.c
@@ -579,7 +579,7 @@ static long ion_sys_cache_sync(struct ion_client *client,
 		m4u_mva_unmap_kernel((unsigned long)param->va,
 				     sync_size, sync_va);
 #else
-		m4u_mva_unmap_kernel((unsigned int)param->va,
+		m4u_mva_unmap_kernel((unsigned long)param->va,
 				     (unsigned int)sync_size, sync_va);
 #endif
 	} else if (ion_need_unmap_flag) {
